[FIX] tests: Give context when installing addons in update mode
Description of the issue:

As of today, if you install & test addons all together in the same moment, the addons that are tested get the module graph of all the addons that are installed before it; but if you install addons and then update & test later (because you don't want to test all addons, but only those) then the test only gets a limited graph that doesn't include all other installed addons.

This shouldn't affect most tests, but if there's an unrelated addon that adds a required field with a default value, or any other SQL constraint, then the test can fail when updating and not when installing.

The fix is pretty simple: Just load addons to update after those that are installed, the same way as it is done currently for addons to install. This way, there's no longer a need to guess addon dependencies and install them separately if you need to test only certain addons.