[FIX] sale: Restore back SO fiscal position
Steps to reproduce the bug:

- Create two fiscal postions fp1 and fp2
- Create a customer C with fp1
- Create a SO for C and set fp2
- Create an invoice for the SO

BUG: The fiscal position of the SO was set to fp1.

Now the fiscal position of the SO is kept when creating an invoice
from a SO (technically, is changed and restored back).

It was solved in c2f15d3, but due to the refactorization introduced in 389bd53,
it was lost again due to merge slots.

This code is very ugly and under-performant (calling 2 times to compute taxes,
an onchange that writes some unneeded data, and then write back again other data),
but I have taken the occasion to at least improve a bit this part not writing again
if the data hasn't changed and doing only one write in that case although both
fields change.