[FIX] website: avoid polluting mass mailing editor
Before this patch, when using the mass mailing editor in `debug=assets` mode after performing some action on the website that added the `force_website_id` key to user's session, the downladed assets would be polluted with the currently-forced website's theme.

To avoid this problem, website-specific assets are only loaded when the user is in a frontend request.

@Tecnativa TT25283