[FIX] website_slides: do not track user without consent
This feature is a little bit unnecessary. It just updates share statistics. The APIs it was calling for Google+, Twitter and LinkedIn were outdated, and the one for Facebook violates GDPR law regarding third party cookies and tracking, and it was already blocked for users who featured tracking blockers in their devices.

The legal and technical problems are not worth the feature, so I'm removing it for the sale of keeping Odoo legal.

@Tecnativa TT26153