[FIX] mail: do not scare away customers
Some customers are scared of clicking on a button that says "Accept & pay" directly from an email.

So, instead of using the online features, they download the PDF, sign it and return it.

It seems legit, due to the text that you can read, so as an UX enhancement, I propose to change that button's text.

@Tecnativa TT27664